backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
python-dotenv==1.2.2
PyJWT==2.11.0
pwdlib[argon2]
email-validator==2.2.0
python-docx==1.1.2
openai==1.76.2
